requirements.txt
# Fast API
fastapi~=0.115.8
uvicorn # ASGI(Asynchronous Server Gateway Interface) 서버로 FastAPI 실행

pydantic-settings~=2.7.1
python-dotenv~=1.0.1

# PostgreSQL 연동
sqlalchemy
psycopg2-binary